Suppress duplicate messaging for PR branches on main repository
More generally, when updating the repository results in two branches
with an identical hash at HEAD, the bot will now only announce the
commits for one of the branches. For the case where a PR is based on a
repository branch, the messages are ordered so that it is the repo
branch that gets announced.  (This could be changed if desired.)

PRs from a forked repository should behave the same as before.